compression: move zlib base to extensions tree
