+++ source/extensions/compression/gzip/common/BUILD
@@ -1,0 +1,19 @@
+licenses(["notice"])  # Apache 2
+
+load(
+    "//bazel:envoy_build_system.bzl",
+    "envoy_cc_library",
+    "envoy_package",
+)
+
+envoy_package()
+
+envoy_cc_library(
+    name = "zlib_base_lib",
+    srcs = ["base.cc"],
+    hdrs = ["base.h"],
+    external_deps = ["zlib"],
+    deps = [
+        "//source/common/buffer:buffer_lib",
+    ],
+)
