+++ source/extensions/compression/gzip/common/base.h
