+++ source/extensions/compression/gzip/compressor/BUILD
@@ -18,7 +18,7 @@ envoy_cc_library(
         "//include/envoy/compression/compressor:compressor_interface",
         "//source/common/buffer:buffer_lib",
         "//source/common/common:assert_lib",
-        "//source/common/common:zlib_base_lib",
+        "//source/extensions/compression/gzip/common:zlib_base_lib",
     ],
 )
